fix runtime error in validation

+++ codes/utils/util.py
@@ -109,7 +109,7 @@ def crop_border(img_list, crop_border):
         return [v[crop_border:-crop_border, crop_border:-crop_border] for v in img_list]
 
 
-def tensor2img(tensor, out_type=np.uint8, min_max=(0, 1)):
+def tensor2img(tensor, out_type=np.uint16, min_max=(0, 1)):
     '''
     Converts a torch Tensor into an image Numpy array
     Input: 4D(B,(3/1),H,W), 3D(C,H,W), or 2D(H,W), any range, RGB channel order
